기본키 타입들을 bigint로 수정

+++ capstone_back/demo/src/main/java/com/example/demo/entity/PW.java
@@ -19,7 +19,7 @@ public class PW {
 
     @Id
     @GeneratedValue
-    private int passwordId;
+    private Long passwordId;
 
     @Column
     private String passwordHash;
