+++ capstone_back/demo/src/main/java/com/example/demo/repository/PasswordRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.entity.PW;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PasswordRepository extends CrudRepository<PW, Integer> {
+public interface PasswordRepository extends CrudRepository<PW, Long> {
 }
